added dependency in Build_gradle to handle pokemondb_impl refactoring in insertPokemon to return Long instead of int DatabaseModule created to provide a singleton of DB

+++ app/src/main/java/com/pokemon_app/domain/repository/IPokemonDbRepository.kt
@@ -5,7 +5,7 @@ import com.pokemon_app.database.PokemonEntity
 import com.pokemon_app.domain.model.Pokemon
 
 interface IPokemonDbRepository {
-    suspend fun insertPokemon(pokemon: PokemonEntity): Int
+    suspend fun insertPokemon(pokemon: PokemonEntity): Long
     suspend fun getPokemon(pokeId: Int): PokemonEntity
     suspend fun deletePokemon(pokemon: PokemonEntity): Int
     suspend fun getAllPokemon(): LiveData<List<PokemonEntity>>
